asi-proj-app/requirements.txt
pandas
numpy
boto3
fastapi
uvicorn
autogluon.tabular==1.3.0
lightgbm
starlette
xgboost
catboost
scikit-learn

matplotlib
seaborn
plotly
shap

tqdm
psutil

torch
torchvision